requirements.txt
addict
scipy
numpy
pillow
pycocotools
